Score is now a real number
Changes all over the board to make sure scores are type double

+++ src/build-game.h
@@ -18,7 +18,7 @@
 #define __INCLUDED_BUILD_GAME_H__
 
 
-struct game* build_new_game(geometry *geo, int difficulty);
+struct game* build_new_game(geometry *geo, double difficulty);
 
 
 #endif
